+++ doc/conf.py
@@ -122,6 +122,7 @@ sphinx_gallery_conf = {
             "plot_FOCISelector_bike_sharing.py",
             "plot_FOCISelector_NN_tie_breaking.py",
             "plot_FOCISelector_average_beats_max.py",
+            "plot_FOCISelector_methods.py",
         ]
     ),
 }